requirements.txt
# Core
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
openai>=1.3.0
python-dotenv>=1.0.0

# Google Cloud
google-cloud-storage>=2.10.0
functions-framework>=3.0.0

# Notifications
python-telegram-bot>=20.7